+++ src/inspect_ai/util/__init__.py
@@ -2,7 +2,7 @@ from inspect_ai._util.trace import trace_action, trace_message
 
 from ._concurrency import concurrency
 from ._console import input_screen
-from ._display import DisplayType, display_type
+from ._display import DisplayType, display_counter, display_type
 from ._panel import InputPanel, input_panel
 from ._resource import resource
 from ._sandbox import (
@@ -31,6 +31,7 @@ __all__ = [
     "ExecResult",
     "concurrency",
     "DisplayType",
+    "display_counter",
     "display_type",
     "InputPanel",
     "input_panel",